[rfc] Define a Mode for local Editor
